Implemented CommandHydrator
