+++ Apollo/CommandSystem/CommandHydrator.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace Apollo.CommandSystem
+{
+    public class CommandHydrator : ICommandHydrator
+    {
+        public void Hydrate(ref ICommand command, object parameters)
+        {
+            var json = JsonConvert.SerializeObject(parameters);
+            JsonConvert.PopulateObject(json, command);
+        }
+    }
+}
